attempt to fix packaging
