+++ tests/CMakeLists.txt
@@ -41,9 +41,9 @@ add_executable(tests_for tests_for.cpp)
 custom_add_test(tests_for)
 
 add_executable(tests_polymorphic_allocator tests_polymorphic_allocator.cpp)
+custom_add_test(tests_polymorphic_allocator)
 
 if (WITH_BOOST)
-  target_compile_definitions(tests_polymorphic_allocator PRIVATE DICE_TEMPLATE_LIBRARY_WITH_BOOST=1)
+  add_executable(tests_offset_ptr_stl_allocator tests_offset_ptr_stl_allocator.cpp)
+  custom_add_test(tests_offset_ptr_stl_allocator)
 endif ()
-
-custom_add_test(tests_polymorphic_allocator)
